requirements.txt
requests==2.28.1
requests-oauthlib==1.3.1
python-dotenv==0.19.2
supabase==0.3.6  # Ensure this matches the version you're using
openai==0.26.5
